change all projects to java 1.8

+++ Qtime-table/Qtime-table-dal/Qtime-table-dal-core/src/main/java/org/xine/qtime/dal/core/services/SubjectService.java
@@ -44,21 +44,22 @@ public class SubjectService {
      * @param s
      *            the s
      * @return the subject
-     * @throws CoreException 
+     * @throws CoreException
      */
     @Transactional
     public Subject save(final Subject s) throws CoreException {
-    	
-    	
-        if ( s == null) {
-          throw  CoreExceptionBuilder.init()
-            					.addType(ExceptionType.SERVICE)
-            					.addSubType(ExceptionSubType.VALIDATION)
-            					.addMessage("Subject Can't be null")
-            					.build();
+
+        if (s == null) {
+            throw CoreExceptionBuilder.init().addType(ExceptionType.SERVICE)
+                    .addSubType(ExceptionSubType.VALIDATION).addMessage("Subject Can't be null")
+                    .build();
         }
-        
-        
+
+        if (s.getName() == null || s.getName().trim().isEmpty()) {
+            throw CoreExceptionBuilder.init().addType(ExceptionType.SERVICE)
+                    .addSubType(ExceptionSubType.VALIDATION).addMessage("Name can't not be null")
+                    .addMessagekey(ValidationErrorsTypes.NAME_EMPTY.toString()).build();
+        }
 
         return this.dao.save(s);
     }
@@ -70,7 +71,19 @@ public class SubjectService {
      * @return the subject
      */
     @Transactional
-    public Subject update(final Subject s) {
+    public Subject update(final Subject s) throws CoreException {
+        if (s == null) {
+            throw CoreExceptionBuilder.init().addType(ExceptionType.SERVICE)
+                    .addSubType(ExceptionSubType.VALIDATION).addMessage("Subject Can't be null")
+                    .build();
+        }
+
+        if (s.getName() == null || s.getName().trim().isEmpty()) {
+            throw CoreExceptionBuilder.init().addType(ExceptionType.SERVICE)
+                    .addSubType(ExceptionSubType.VALIDATION).addMessage("Name can't not be null")
+                    .addMessagekey(ValidationErrorsTypes.NAME_EMPTY.toString()).build();
+        }
+
         return this.dao.update(s);
     }
 
@@ -78,9 +91,23 @@ public class SubjectService {
      * Delete.
      * @param s
      *            the s
+     * @throws CoreException
      */
     @Transactional
-    public void delete(final Subject s) {
+    public void delete(final Subject s) throws CoreException {
+
+        if (s == null) {
+            throw CoreExceptionBuilder.init().addType(ExceptionType.SERVICE)
+                    .addSubType(ExceptionSubType.VALIDATION).addMessage("Subject Can't be null")
+                    .build();
+        }
+
+        if (s.getId() == null || s.getId().longValue() <= 0L) {
+            throw CoreExceptionBuilder.init().addType(ExceptionType.SERVICE)
+            .addSubType(ExceptionSubType.VALIDATION).addMessage("ID is invalalid value")
+            .addMessagekey(ValidationErrorsTypes.INVALID_ID.toString()).build();
+        }
+
         this.dao.delete(s);
     }
